+++ react_frontend/src/App.js
@@ -1,13 +1,15 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import HomeLandingPage from './pages/HomeLandingPage';
-import PatientLoginPage from './pages/PatientLoginPage';
-import PatientHomePage from './pages/PatientHomePage';
-import PatientReportingLandingPage from './pages/PatientReportingLandingPage';
-import InfoPage from './pages/InfoPage';
-import GeneralInstructionsPage from './pages/GeneralInstructionsPage';
-import LightingCalibration from './pages/LightingCalibration';
-import GazeCalibration from './pages/GazeCalibration';
+import PatientLoginPage from './pages/patient/PatientLoginPage';
+import PatientHomePage from './pages/patient/PatientHomePage';
+import PatientReportingLandingPage from './pages/patient/PatientReportingLandingPage';
+import InfoPage from './pages/patient/InfoPage';
+import GeneralInstructionsPage from './pages/patient/GeneralInstructionsPage';
+import LightingCalibration from './pages/patient/LightingCalibration';
+import GazeCalibration from './pages/patient/GazeCalibration';
+import ProsaccadeInstructions from './pages/games/prosaccade/ProsaccadeInstructions';
+import ProsaccadeGame from './pages/games/prosaccade/ProsaccadeGame';
 import './App.css'; // Global styles
 
 const App = () => {
@@ -22,6 +24,8 @@ const App = () => {
         <Route path="/general-instructions" element={<GeneralInstructionsPage />} />
         <Route path="/lighting-calibration" element={<LightingCalibration />} />
         <Route path="/gaze-calibration" element={<GazeCalibration />} />
+        <Route path="/prosaccade-instructions" element={<ProsaccadeInstructions />} />
+        <Route path="/prosaccade-game" element={<ProsaccadeGame />} />
       </Routes>
     </Router>
   );
