+++ react_frontend/src/pages/games/prosaccade/ProsaccadeGame.css
@@ -1,0 +1,21 @@
+.prosaccade-game-container {
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: #fff3e3;
+}
+
+.prosaccade-game-container h1 {
+  font-size: 42px;
+  font-weight: bold;
+  margin-bottom: 20px;
+}
+
+.prosaccade-game-container p {
+  font-size: 18px;
+  line-height: 1.6;
+  margin-bottom: 20px;
+}
