+++ react_frontend/src/pages/games/prosaccade/ProsaccadeGame.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './ProsaccadeGame.css';
+
+const ProsaccadeGame = () => {
+  return (
+    <div className="prosaccade-game-container">
+      <h1>Prosaccade Game</h1>
+      <p>Game content goes here...</p>
+    </div>
+  );
+};
+
+export default ProsaccadeGame;
