+++ react_frontend/src/pages/games/prosaccade/ProsaccadeInstructions.css
@@ -1,0 +1,40 @@
+.prosaccade-instructions-container {
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: #fff3e3;
+}
+
+.prosaccade-instructions-container h1 {
+  font-size: 42px;
+  font-weight: bold;
+  margin-bottom: 20px;
+}
+
+.prosaccade-instructions-container p {
+  font-size: 18px;
+  line-height: 1.6;
+  margin-bottom: 20px;
+}
+
+.next-button {
+  font-weight: bold;
+  width: auto;
+  color: white;
+  background-color: #576a7f;
+  border: none;
+  cursor: pointer;
+  transition: background-color 0.3s;
+  border-radius: 50px;
+  text-decoration: none;
+  padding: 14px 24px;
+  font-size: 17px;
+  margin-top: 20px;
+}
+
+.next-button:hover {
+  background-color: #8496aa;
+}
