+++ react_frontend/src/pages/patient/GeneralInstructionsPage.css
