+++ react_frontend/src/pages/patient/GeneralInstructionsPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import titleImage from '../assets/title.svg';
+import titleImage from '../../assets/title.svg';
 import './GeneralInstructionsPage.css';
 
 const GeneralInstructionsPage = () => {
