+++ react_frontend/src/pages/patient/InfoPage.css
