+++ react_frontend/src/pages/patient/InfoPage.js
