+++ react_frontend/src/pages/patient/LightingCalibration.css
