+++ react_frontend/src/pages/patient/LightingCalibration.js
